change app2.py to app.py
